+++ internal/cmd/refresh.go
@@ -29,13 +29,17 @@ var (
 				return err
 			}
 
+			fmt.Println("Running...")
+
 			// run the script
-			_, err = script.Run(true, `sh /tmp/refresh.sh `+Version)
+			output, err := script.Run(true, `bash /tmp/refresh.sh `+Version)
 			if err != nil {
 				return err
 			}
 
-			fmt.Println("Refreshed the templates and configs for the machine to ", Version)
+			fmt.Println(output)
+
+			fmt.Println("Refreshed the templates and configs for the machine to", Version)
 
 			return nil
 		},